refactor: moved schemas to separate files
