+++ energetica/simulate.py
@@ -3,16 +3,15 @@
 from __future__ import annotations
 
 import cProfile
-from datetime import datetime
 import pstats
 from time import sleep
-from typing import Annotated, Any, Literal, cast
+from typing import Any, cast
 
-from pydantic import BaseModel, Field
 import requests
 
 from energetica.database.player import Player
 from energetica.globals import engine
+from energetica.schemas.simulate import Action
 from energetica.utils.tick_execution import tick
 
 base_url: str | None = None
@@ -179,47 +178,3 @@ def _simulate(
     else:
         return True
     return False
-
-
-class InitEngineAction(BaseModel):
-    instance_uuid: str
-    env: Literal["dev", "prod"]
-    clock_time: int
-    in_game_seconds_per_tick: int
-    action_type: Literal["init_engine"]
-    random_seed: int
-    start_date: datetime
-    disable_signups: bool
-
-
-class CreateUserAction(BaseModel):
-    timestamp: datetime
-    ip: str
-    action_type: Literal["create_user"]
-    player_id: int
-    username: str
-    pw_hash: str
-
-
-class TickAction(BaseModel):
-    timestamp: datetime
-    action_type: Literal["tick"]
-    total_t: int
-    elapsed: float
-
-
-class RequestAction(BaseModel):
-    timestamp: datetime
-    elapsed: float
-    ip: str
-    action_type: Literal["request"]
-    player_id: int
-    request: dict
-    response: dict
-
-    # "request": {"endpoint": "/api/v1/map/252:settle", "content_type": "application/json", "content": {}},
-    # "response": {"status_code": 204, "content_type": "application/json", "content": "unparsable"},
-
-
-ActionUnionType = InitEngineAction | CreateUserAction | TickAction | RequestAction
-Action = Annotated[ActionUnionType, Field(discriminator="action_type")]
